refactor: hide fakes package so it does not show up on godoc
