+++ pkg/notes/internal/fakes/trainstat_client.go
